# Pinta 15% faster