Some runtimes:
Previous: 12 ms, 11.7 ms, 11.6 ms, 11.5 ms, 11.6 ms
New:    10.1 ms, 9.85 ms, 9.9 ms, 9.93 ms, 9.79 ms
Result: Approx. 2 ms, 15% faster

Maakt gebruik van het feit dat de "pinta ratio functie" convex is
Waarden ver links van het minimum hoeven dus niet bekeken te worden